@@ -4,7 +4,8 @@ description: Demonstrates how to use the adv_camera plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.2 <3.0.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
