@@ -6,59 +6,56 @@ import 'package:flutter/material.dart';
 class CameraApp extends StatefulWidget {
   final String id;
 
-  const CameraApp({Key key, this.id}) : super(key: key);
+  const CameraApp({Key? key, required this.id}) : super(key: key);
 
   @override
   _CameraAppState createState() => _CameraAppState();
 }
 
 class _CameraAppState extends State<CameraApp> {
-  List<String> pictureSizes = [];
-  String imagePath;
+  List<String> pictureSizes = <String>[];
+  String? imagePath;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      // appBar: AppBar(
-      //   title: const Text('AdvCamera Example'),
-      // ),
+      appBar: AppBar(
+        title: const Text('AdvCamera Example'),
+      ),
       body: SafeArea(
         child: Stack(
           children: [
             Column(
               children: [
-                // Padding(
-                //   padding: EdgeInsets.symmetric(vertical: 16),
-                //   child: Column(
-                //     crossAxisAlignment: CrossAxisAlignment.start,
-                //     children: [
-                //       buildFlashSettings(context),
-                //       Container(
-                //         margin: EdgeInsets.symmetric(
-                //           horizontal: 16,
-                //         ).copyWith(bottom: 16),
-                //         height: 1,
-                //         width: double.infinity,
-                //         color: Colors.grey,
-                //       ),
-                //       buildRatioSettings(context),
-                //       if (this.pictureSizes.isNotEmpty)
-                //         Container(
-                //           margin: EdgeInsets.symmetric(
-                //             horizontal: 16,
-                //           ).copyWith(bottom: 16),
-                //           height: 1,
-                //           width: double.infinity,
-                //           color: Colors.grey,
-                //         ),
-                //       if (this.pictureSizes.isNotEmpty)
-                //         buildImageOutputSettings(context),
-                //     ],
-                //   ),
-                // ),
-        // D/handleFocus(23032): event.getX() => 1011.74994
-        // D/handleFocus(23032): event.getY() => 1289.25
-        //         1050.75 x 1288.5
+                Padding(
+                  padding: EdgeInsets.symmetric(vertical: 16),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      buildFlashSettings(context),
+                      Container(
+                        margin: EdgeInsets.symmetric(
+                          horizontal: 16,
+                        ).copyWith(bottom: 16),
+                        height: 1,
+                        width: double.infinity,
+                        color: Colors.grey,
+                      ),
+                      buildRatioSettings(context),
+                      if (this.pictureSizes.isNotEmpty)
+                        Container(
+                          margin: EdgeInsets.symmetric(
+                            horizontal: 16,
+                          ).copyWith(bottom: 16),
+                          height: 1,
+                          width: double.infinity,
+                          color: Colors.grey,
+                        ),
+                      if (this.pictureSizes.isNotEmpty)
+                        buildImageOutputSettings(context),
+                    ],
+                  ),
+                ),
                 Expanded(
                   child: AdvCamera(
                     initialCameraType: CameraType.rear,
@@ -76,52 +73,52 @@ class _CameraAppState extends State<CameraApp> {
                 ),
               ],
             ),
-            // Positioned(
-            //   bottom: 16.0,
-            //   left: 16.0,
-            //   child: imagePath != null
-            //       ? Container(
-            //           width: 100.0,
-            //           height: 100.0,
-            //           child: Image.file(File(imagePath)),
-            //         )
-            //       : Icon(Icons.image),
-            // )
+            Positioned(
+              bottom: 16.0,
+              left: 16.0,
+              child: imagePath != null
+                  ? Container(
+                      width: 100.0,
+                      height: 100.0,
+                      child: Image.file(File(imagePath!)),
+                    )
+                  : Icon(Icons.image),
+            )
           ],
         ),
       ),
-      // floatingActionButton: Column(
-      //   crossAxisAlignment: CrossAxisAlignment.end,
-      //   mainAxisSize: MainAxisSize.min,
-      //   children: [
-      //     FloatingActionButton(
-      //       heroTag: "switch",
-      //       child: Icon(Icons.switch_camera),
-      //       onPressed: () async {
-      //         await cameraController.switchCamera();
-      //       },
-      //     ),
-      //     Container(height: 16.0),
-      //     FloatingActionButton(
-      //       heroTag: "capture",
-      //       child: Icon(Icons.camera),
-      //       onPressed: () {
-      //         cameraController.captureImage();
-      //       },
-      //     ),
-      //   ],
-      // ),
+      floatingActionButton: Column(
+        crossAxisAlignment: CrossAxisAlignment.end,
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          FloatingActionButton(
+            heroTag: "switch",
+            child: Icon(Icons.switch_camera),
+            onPressed: () async {
+              await cameraController!.switchCamera();
+            },
+          ),
+          Container(height: 16.0),
+          FloatingActionButton(
+            heroTag: "capture",
+            child: Icon(Icons.camera),
+            onPressed: () {
+              cameraController!.captureImage();
+            },
+          ),
+        ],
+      ),
     );
   }
 
-  AdvCameraController cameraController;
+  AdvCameraController? cameraController;
 
   _onCameraCreated(AdvCameraController controller) {
     this.cameraController = controller;
 
-    this.cameraController.getPictureSizes().then((pictureSizes) {
+    this.cameraController!.getPictureSizes().then((pictureSizes) {
       setState(() {
-        this.pictureSizes = pictureSizes;
+        this.pictureSizes = pictureSizes ?? <String>[];
       });
     });
   }
@@ -139,28 +136,28 @@ class _CameraAppState extends State<CameraApp> {
           child: Container(
             child: Row(
               children: [
-                FlatButton(
+                TextButton(
                   child: Text("Auto"),
                   onPressed: () {
-                    cameraController.setFlashType(FlashType.auto);
+                    cameraController!.setFlashType(FlashType.auto);
                   },
                 ),
-                FlatButton(
+                TextButton(
                   child: Text("On"),
                   onPressed: () {
-                    cameraController.setFlashType(FlashType.on);
+                    cameraController!.setFlashType(FlashType.on);
                   },
                 ),
-                FlatButton(
+                TextButton(
                   child: Text("Off"),
                   onPressed: () {
-                    cameraController.setFlashType(FlashType.off);
+                    cameraController!.setFlashType(FlashType.off);
                   },
                 ),
-                FlatButton(
+                TextButton(
                   child: Text("Torch"),
                   onPressed: () {
-                    cameraController.setFlashType(FlashType.torch);
+                    cameraController!.setFlashType(FlashType.torch);
                   },
                 ),
               ],
@@ -184,33 +181,34 @@ class _CameraAppState extends State<CameraApp> {
           child: Container(
             child: Row(
               children: [
-                FlatButton(
+                TextButton(
                   child: Text(Platform.isAndroid ? "1:1" : "Low"),
                   onPressed: () {
-                    cameraController.setPreviewRatio(CameraPreviewRatio.r1);
-                    cameraController.setSessionPreset(CameraSessionPreset.low);
+                    cameraController!.setPreviewRatio(CameraPreviewRatio.r1);
+                    cameraController!.setSessionPreset(CameraSessionPreset.low);
                   },
                 ),
-                FlatButton(
+                TextButton(
                   child: Text(Platform.isAndroid ? "4:3" : "Medium"),
                   onPressed: () {
-                    cameraController.setPreviewRatio(CameraPreviewRatio.r4_3);
-                    cameraController
+                    cameraController!.setPreviewRatio(CameraPreviewRatio.r4_3);
+                    cameraController!
                         .setSessionPreset(CameraSessionPreset.medium);
                   },
                 ),
-                FlatButton(
+                TextButton(
                   child: Text(Platform.isAndroid ? "11:9" : "High"),
                   onPressed: () {
-                    cameraController.setPreviewRatio(CameraPreviewRatio.r11_9);
-                    cameraController.setSessionPreset(CameraSessionPreset.high);
+                    cameraController!.setPreviewRatio(CameraPreviewRatio.r11_9);
+                    cameraController!
+                        .setSessionPreset(CameraSessionPreset.high);
                   },
                 ),
-                FlatButton(
+                TextButton(
                   child: Text(Platform.isAndroid ? "16:9" : "Best"),
                   onPressed: () {
-                    cameraController.setPreviewRatio(CameraPreviewRatio.r16_9);
-                    cameraController
+                    cameraController!.setPreviewRatio(CameraPreviewRatio.r16_9);
+                    cameraController!
                         .setSessionPreset(CameraSessionPreset.photo);
                   },
                 ),
@@ -234,14 +232,17 @@ class _CameraAppState extends State<CameraApp> {
           scrollDirection: Axis.horizontal,
           child: Row(
             children: this.pictureSizes.map((pictureSize) {
-              return FlatButton(
+              return TextButton(
                 child: Text(pictureSize),
                 onPressed: () {
-                  cameraController.setPictureSize(
-                      int.tryParse(
-                          pictureSize.substring(0, pictureSize.indexOf(":"))),
+                  cameraController!.setPictureSize(
                       int.tryParse(pictureSize.substring(
-                          pictureSize.indexOf(":") + 1, pictureSize.length)));
+                              0, pictureSize.indexOf(":"))) ??
+                          0,
+                      int.tryParse(pictureSize.substring(
+                              pictureSize.indexOf(":") + 1,
+                              pictureSize.length)) ??
+                          0);
                 },
               );
             }).toList(),
